+++ csatar-plugins/forms/traits/AjaxControllerSimple.php
@@ -706,13 +706,18 @@ trait AjaxControllerSimple {
             }
 
             $key = isset($definition['key']) ? $definition['key'] : $name . '_id';
-            $data[$key] = (int) $data[$name];
+            if (isset($definition['keyType'])) {
+                $data[$key] = $data[$name];
+                settype($data[$key], $definition['keyType']);
+            } else {
+                $data[$key] = (int) $data[$name];
+            }
 //            unset($data[$name]);
         }
 
         // Resolve belongsToMany relations
         foreach($record->belongsToMany as $relationName => $definition) {
-            if (!isset($data[$relationName]) || $data[$relationName] =='' || !isset($data[$relationName]['pivot'])) {
+            if (!isset($data[$relationName]) || $data[$relationName] =='') {
                 continue;
             }
 
@@ -771,14 +776,6 @@ trait AjaxControllerSimple {
 
             // resolve other relations
             $record->commitDeferred($this->sessionKey);
-        }
-
-        // Resolve belongsToMany relations
-        foreach($record->belongsToMany as $name => $definition) {
-            if (!isset($data[$name]) || $data[$name] =='' || !isset($data[$name]['pivot'])) {
-                continue;
-            }
-            $record->$name()->sync($data[$name]);
         }
 
         if (!$record->update($data) && !$isNew) {