+++ csatar-plugins/knowledgerepository/updates/SeederData.php
@@ -3,10 +3,11 @@
 use Csatar\KnowledgeRepository\Models\AccidentRiskLevel;
 use Csatar\KnowledgeRepository\Models\GameDevelopmentGoal;
 use Csatar\KnowledgeRepository\Models\Tool;
-use Csatar\KnowledgeRepository\Models\HeadCount;
+use Csatar\KnowledgeRepository\Models\Headcount;
 use Csatar\KnowledgeRepository\Models\Duration;
 use Csatar\KnowledgeRepository\Models\Location;
 use Csatar\KnowledgeRepository\Models\GameType;
+use Csatar\Forms\Models\Form;
 use Db;
 use Seeder;
 
@@ -260,6 +261,12 @@ class SeederData extends Seeder
                 'sort_order' => 6,
             ],
         ],
+        'forms' => [
+            [
+                'title' => 'Játék',
+                'model' => 'Csatar\KnowledgeRepository\Models\Game',
+            ],
+        ],
     ];
 
     public function run()
@@ -293,7 +300,7 @@ class SeederData extends Seeder
 
         // Head Counts
         foreach ($this::DATA['headCounts'] as $headCountData) {
-            $headCount = HeadCount::firstOrNew([
+            $headCount = Headcount::firstOrNew([
                 'description' => $headCountData['description'],
             ]);
             $headCount->min = $headCountData['min'];
@@ -329,6 +336,12 @@ class SeederData extends Seeder
             ]);
             $gameType->sort_order = $gameTypeData['sort_order'];
             $gameType->save();
+        }
+
+        // Forms
+        foreach ($this::DATA['forms'] as $formData) {
+            $form = Form::firstOrNew($formData);
+            $form->save();
         }
     }
 }